Add keeper reward plus coll distributor (#1994)
* wip

* Introduce collateralLiquidateRewardRatioD18

* add reward distributor

* wip

* wip

* duh... make it build again

* some fixes

* remove storage checking on distributor auxiliar

* Introduce CollateralConfigurationModule

* include in cannonfile

* move to collateralConfigurationModule

* comments and TODOs

* add some comments

* wip

* add collateral rewards distributor to bootstrap

* Add version to PerpsRewardDistributor

* some fixes while adding tests

* wip test

* wip compilation fixes after rebase

* make it compile

* (wip) fix tests

* lint

* update old dependencies

---------

+++ markets/perps-market/contracts/modules/PerpsAccountModule.sol
@@ -99,7 +99,7 @@ contract PerpsAccountModule is IPerpsAccountModule {
     /**
      * @inheritdoc IPerpsAccountModule
      */
-    function totalCollateralValue(uint128 accountId) external view override returns (uint) {
+    function totalCollateralValue(uint128 accountId) external view override returns (uint256) {
         return
             PerpsAccount.load(accountId).getTotalCollateralValue(
                 PerpsPrice.Tolerance.DEFAULT,
